+++ HISTORY.md
@@ -6,3 +6,15 @@
 ## 0.1.1 (2023-09-11)
 ### Corrected
 - Function name from compute_ecdf to multivariate_ecdf
+
+## 0.1.2 (2023-09-11)
+### Enhanced
+- Added a new parameter verbose in multivariate ECDF function
+
+## 0.1.3 (2023-09-11)
+### Corrected
+- Removed unecessary docstrings from code
+
+## 0.1.4 (2023-09-11)
+### Fixed
+- Resolved issues with special characters in the input Pandas DataFrames